refactor: Dedicated account menu component
